+++ src/main/scala/wdlTools/eval/Eval.scala
@@ -1,10 +1,11 @@
 package wdlTools.eval
 
+import dx.util.{Bindings, EvalPaths, FileSource, FileSourceResolver, LocalFileSource, Logger}
+import dx.util.CollectionUtils.IterableOnceExtensions
 import wdlTools.eval.WdlValues._
 import wdlTools.syntax.{SourceLocation, WdlVersion}
 import wdlTools.types.ExprState.ExprState
 import wdlTools.types.{ExprState, TypeUtils, WdlTypes, TypedAbstractSyntax => TAT}
-import dx.util.{Bindings, EvalPaths, FileSource, FileSourceResolver, LocalFileSource, Logger}
 
 import scala.collection.immutable.TreeSeqMap
 
@@ -34,8 +35,23 @@ case class Eval(paths: EvalPaths,
     }
   }
 
+  /**
+    * The context contains variable bindings that can be referenced when evaluating an
+    * expression, and also an expression state, which determies which evaluation rules
+    * apply.
+    *
+    * Bindings may contain fully-qualified variable names, such as `a.b.c`, where
+    * `a` is a Call, Pair, or Struct. In practice, this can be used by a runtime engine
+    * to simplify call inputs/outputs - e.g. if task `bar` refers to the output of task `foo`
+    * (say `foo.result`), rather than having to make the `foo` call object (or some serialized
+    * version thereof) avaialble to `bar`, it can just  expose an output parameter `foo.result`.
+    * @param bindings variable bindings
+    * @param exprState current expression state
+    */
   private case class Context(bindings: Bindings[String, V] = WdlValueBindings.empty,
                              exprState: ExprState = ExprState.Start) {
+
+    lazy val hasFullyQualifiedBindings: Boolean = bindings.keySet.exists(_.contains("."))
 
     def apply(name: String): V = {
       bindings
@@ -46,13 +62,14 @@ case class Eval(paths: EvalPaths,
     }
 
     def advanceState(condition: Option[ExprState] = None): Context = {
-      if (condition.exists(exprState >= _)) {
-        throw new Exception(s"Context in an invalid state ${exprState} >= ${condition.get}")
-      }
       val newState = exprState match {
-        case ExprState.Start    => ExprState.InString
-        case ExprState.InString => ExprState.InPlaceholder
-        case _                  => throw new Exception(s"Cannot advance state from ${exprState}")
+        case ExprState.Start         => ExprState.InString
+        case ExprState.InString      => ExprState.InPlaceholder
+        case ExprState.InPlaceholder => ExprState.InString // nested string
+        case _                       => throw new Exception(s"Cannot advance state from ${exprState}")
+      }
+      if (condition.exists(_ < newState)) {
+        throw new Exception(s"Context in an invalid state ${newState} >= ${condition.get}")
       }
       copy(exprState = newState)
     }
@@ -98,6 +115,83 @@ case class Eval(paths: EvalPaths,
     }
   }
 
+  /**
+    * Resolves a fully-qualified identifier.
+    * @param expr the LHS expression to resolve
+    * @param fieldName the RHS field name
+    * @param ctx the Context
+    * @return if the expression resolves to a fully-qualfied name and that name
+    *         is in `ctx`, returns the bound value, otherwise None
+    * @example
+    * Given a declaration:
+    *   `Pair[Pair[String, Int], File] foo`
+    * and an expression:
+    *   `foo.left.right`
+    * there are two possible scenarios:
+    *   1. The context contains binding `foo -> V_Pair(...)`,
+    *      in which case the expression is resolved recursively by `apply`,
+    *   2. The context contains binding `foo.left.right -> V_Int(...)`,
+    *      in which case the entire expression needs to be resolved at once
+    *      to derive the fully-qualified name to look up in the context.
+    */
+  private def resolveFullyQualifedIdentifier(expr: TAT.Expr,
+                                             fieldName: String,
+                                             ctx: Context): Option[V] = {
+    // tests whether the given field name is legal for the given type
+    def canResolveField(t: WdlTypes.T, field: String): Boolean = {
+      TypeUtils.unwrapOptional(t) match {
+        case _: WdlTypes.T_Pair if Set("left", "right").contains(field) => true
+        case WdlTypes.T_Struct(_, fields) if fields.contains(field)     => true
+        case WdlTypes.T_Call(_, outputs) if outputs.contains(field)     => true
+        // there's no way to guarantee the resolution is valid for these types:
+        case WdlTypes.T_Object => true
+        case WdlTypes.T_Any    => true
+        case _                 => false
+      }
+    }
+
+    // builds a Vector of name components
+    def inner(innerExpr: TAT.Expr, innerFieldName: String): Option[Vector[String]] = {
+      innerExpr match {
+        case TAT.ExprIdentifier(id, wdlType, _) if canResolveField(wdlType, innerFieldName) =>
+          Some(Vector(id, innerFieldName))
+        case TAT.ExprGetName(e, fieldName, _, _) =>
+          inner(e, fieldName).map(v => v :+ innerFieldName)
+        case _ => None
+      }
+    }
+
+    // Given a Vector of name components, moving from right to left, split the
+    // Vector at each index - the left-side will (when concatenated with '.'
+    // separator) give a fully-qualified name that we look up in the context. If
+    // it resolves to a value, then we recursively apply the field names in the
+    // right side to get the final value.
+    //
+    // For example, say that the context contains `a.b -> Pair("x", "y")`. If we
+    // try to resolve `a.b.left`, it is not in the context, so then we go to `a.b`,
+    // get the value, and use it to resolve `left`, with the final result being "x".
+    inner(expr, fieldName).flatMap { v =>
+      v.size
+        .to(1, -1)
+        .iterator
+        .map(v.splitAt)
+        .collectFirstDefined {
+          case (path, fields) =>
+            ctx.bindings.get(path.mkString(".")).flatMap { value =>
+              fields.foldLeft(Option(value)) {
+                case (Some(v), name) =>
+                  try {
+                    Some(exprGetName(v, name, expr.loc))
+                  } catch {
+                    case _: Throwable => None
+                  }
+                case _ => None
+              }
+            }
+        }
+    }
+  }
+
   // within an interpolation, null/None renders as empty string
   private def interpolationValueToString(V: V, loc: SourceLocation): String = {
     V match {
@@ -124,6 +218,14 @@ case class Eval(paths: EvalPaths,
                     ctx: Context,
                     validate: (V, SourceLocation) => Unit = validateDefault): V = {
     def inner(nestedExpr: TAT.Expr, nestedCtx: Context): V = {
+      val updatedCtx = nestedExpr match {
+        case _: TAT.ValueString                             => nestedCtx
+        case _ if nestedCtx.exprState == ExprState.InString =>
+          // if we're in a string and the current expression is anything
+          // other than ValueString, that means we're in a placeholder
+          nestedCtx.advanceState()
+        case _ => nestedCtx
+      }
       val v = nestedExpr match {
         case _: TAT.ValueNull      => V_Null
         case _: TAT.ValueNone      => V_Null
@@ -136,14 +238,14 @@ case class Eval(paths: EvalPaths,
 
         // complex types
         case TAT.ExprPair(left, right, _, _) =>
-          V_Pair(inner(left, nestedCtx), inner(right, nestedCtx))
+          V_Pair(inner(left, updatedCtx), inner(right, updatedCtx))
         case TAT.ExprArray(value, _, _) =>
           V_Array(value.map { x =>
-            inner(x, nestedCtx)
+            inner(x, updatedCtx)
           })
         case TAT.ExprMap(value, _, _) =>
           V_Map(value.map {
-            case (k, v) => inner(k, nestedCtx) -> inner(v, nestedCtx)
+            case (k, v) => inner(k, updatedCtx) -> inner(v, updatedCtx)
           })
         case TAT.ExprObject(value, _, _) =>
           V_Object(
@@ -151,7 +253,7 @@ case class Eval(paths: EvalPaths,
                 .map {
                   case (k, v) =>
                     // an object literal key can be a string or identifier
-                    val key = inner(k, nestedCtx) match {
+                    val key = inner(k, updatedCtx) match {
                       case V_String(s) => s
                       case _ =>
                         throw new EvalException(
@@ -159,13 +261,13 @@ case class Eval(paths: EvalPaths,
                             expr.loc
                         )
                     }
-                    key -> inner(v, nestedCtx)
+                    key -> inner(v, updatedCtx)
                 }
                 .to(TreeSeqMap)
           )
 
-        case TAT.ExprIdentifier(id, _, _) if nestedCtx.bindings.contains(id) =>
-          nestedCtx.bindings(id)
+        case TAT.ExprIdentifier(id, _, _) if updatedCtx.bindings.contains(id) =>
+          updatedCtx.bindings(id)
         case TAT.ExprIdentifier(id, _, _) =>
           throw new EvalException(s"identifier ${id} not found")
 
@@ -173,17 +275,17 @@ case class Eval(paths: EvalPaths,
         case TAT.ExprCompoundString(value, _, _) =>
           // concatenate an array of strings inside an expression/command block
           val strArray: Vector[String] = value.map { expr =>
-            val v = inner(expr, nestedCtx.advanceState(Some(ExprState.InString)))
+            val v = inner(expr, updatedCtx.advanceState(Some(ExprState.InString)))
             interpolationValueToString(v, expr.loc)
           }
           V_String(strArray.mkString(""))
         case TAT.ExprPlaceholderCondition(t, f, boolExpr, _, _) =>
           // ~{true="--yes" false="--no" boolean_value}
-          inner(boolExpr, nestedCtx.advanceState(Some(ExprState.InPlaceholder))) match {
+          inner(boolExpr, updatedCtx.advanceState(Some(ExprState.InPlaceholder))) match {
             case V_Boolean(true) =>
-              inner(t, nestedCtx.advanceState(Some(ExprState.InPlaceholder)))
+              inner(t, updatedCtx.advanceState(Some(ExprState.InPlaceholder)))
             case V_Boolean(false) =>
-              inner(f, nestedCtx.advanceState(Some(ExprState.InPlaceholder)))
+              inner(f, updatedCtx.advanceState(Some(ExprState.InPlaceholder)))
             case other =>
               throw new EvalException(
                   s"invalid boolean value ${other}, should be a boolean",
@@ -192,17 +294,17 @@ case class Eval(paths: EvalPaths,
           }
         case TAT.ExprPlaceholderDefault(defaultVal, optVal, _, _) =>
           // ~{default="foo" optional_value}
-          inner(optVal, nestedCtx.advanceState(Some(ExprState.InPlaceholder))) match {
-            case V_Null => inner(defaultVal, nestedCtx.advanceState(Some(ExprState.InPlaceholder)))
+          inner(optVal, updatedCtx.advanceState(Some(ExprState.InPlaceholder))) match {
+            case V_Null => inner(defaultVal, updatedCtx.advanceState(Some(ExprState.InPlaceholder)))
             case other  => other
           }
         case TAT.ExprPlaceholderSep(sep: TAT.Expr, arrayVal: TAT.Expr, _, _) =>
           // ~{sep=", " array_value}
           val sepString = interpolationValueToString(
-              inner(sep, nestedCtx.advanceState(Some(ExprState.InPlaceholder))),
+              inner(sep, updatedCtx.advanceState(Some(ExprState.InPlaceholder))),
               sep.loc
           )
-          inner(arrayVal, nestedCtx.advanceState(Some(ExprState.InPlaceholder))) match {
+          inner(arrayVal, updatedCtx.advanceState(Some(ExprState.InPlaceholder))) match {
             case V_Array(array) =>
               val elements: Vector[String] = array.map(interpolationValueToString(_, expr.loc))
               V_String(elements.mkString(sepString))
@@ -212,17 +314,17 @@ case class Eval(paths: EvalPaths,
 
         case TAT.ExprIfThenElse(cond, tBranch, fBranch, _, loc) =>
           // if (x == 1) then "Sunday" else "Weekday"
-          inner(cond, nestedCtx) match {
-            case V_Boolean(true)  => inner(tBranch, nestedCtx)
-            case V_Boolean(false) => inner(fBranch, nestedCtx)
+          inner(cond, updatedCtx) match {
+            case V_Boolean(true)  => inner(tBranch, updatedCtx)
+            case V_Boolean(false) => inner(fBranch, updatedCtx)
             case _ =>
               throw new EvalException(s"condition is not boolean", loc)
           }
 
         case TAT.ExprAt(collection, index, _, loc) =>
           // Access an array element at [index: Int] or map value at [key: K]
-          val collectionVal = inner(collection, nestedCtx)
-          val indexVal = inner(index, nestedCtx)
+          val collectionVal = inner(collection, updatedCtx)
+          val indexVal = inner(index, updatedCtx)
           (collectionVal, indexVal) match {
             case (V_Array(av), V_Int(n)) if n < av.size =>
               av(n.toInt)
@@ -251,45 +353,28 @@ case class Eval(paths: EvalPaths,
               )
           }
 
-        case TAT.ExprGetName(TAT.ExprIdentifier(id, idType, _), fieldName, _, _)
-            if nestedCtx.bindings.contains(s"${id}.${fieldName}") =>
-          // These cases are a bit of a hack - they enable resolution of a fully-qualified name
-          // when the LHS is an identifier, rather than having to first evaluate the LHS expression.
-          // In practice, this can be used by a runtime engine to simplify call inputs/outputs - e.g.
-          // if task Bar refers to the output of task Foo (say Foo.result), rather than having to make
-          // the Foo call object (or some serialized version thereof) avaialble to Bar, it can just
-          // expose an output parameter "Foo.result".
-          val fqnResolutionAllowed = TypeUtils.unwrapOptional(idType) match {
-            case _: WdlTypes.T_Pair if Set("left", "right").contains(fieldName) => true
-            case WdlTypes.T_Struct(_, members) if members.contains(fieldName)   => true
-            case WdlTypes.T_Call(_, output) if output.contains(fieldName)       => true
-            // there's no way to guarantee the resolution is valid for these types:
-            case WdlTypes.T_Object => true
-            case WdlTypes.T_Any    => true
-            case _                 => false
-          }
-          if (fqnResolutionAllowed) {
-            nestedCtx.bindings(s"${id}.${fieldName}")
-          } else {
-            throw new EvalException(
-                s"""'${id}.${fieldName}' is present in the evaluation context, but fully-qualified name 
-                   |resolution is not allowed for identifier of type ${idType}""".stripMargin
-                  .replaceAll("\n", " ")
-            )
-          }
-
         case TAT.ExprGetName(e: TAT.Expr, fieldName, _, loc) =>
-          // Evaluate the expression, then access the field
-          val ev = inner(e, nestedCtx)
-          exprGetName(ev, fieldName, loc)
+          Option
+            .when(updatedCtx.hasFullyQualifiedBindings) {
+              // if the context has fully-qualified identifiers, try to
+              // resolve the expression as a identifier with chained
+              // field accesses
+              resolveFullyQualifedIdentifier(e, fieldName, updatedCtx)
+            }
+            .flatten
+            .getOrElse {
+              // evaluate the LHS expression, then access the RHS field
+              val ev = inner(e, updatedCtx)
+              exprGetName(ev, fieldName, loc)
+            }
 
         case TAT.ExprApply(funcName, _, elements, _, loc) =>
           // Apply a standard library function (including built-in operators)
           // to arguments. For example:
           //   1 + 1
           //   read_int("4")
-          val funcArgs = elements.map(e => inner(e, nestedCtx))
-          standardLibrary.call(funcName, funcArgs, loc, nestedCtx.exprState)
+          val funcArgs = elements.map(e => inner(e, updatedCtx))
+          standardLibrary.call(funcName, funcArgs, loc, updatedCtx.exprState)
 
         case other =>
           throw new Exception(s"sanity: expression ${other} not implemented")