+++ src/main/scala/wdlTools/eval/WdlValues.scala
@@ -34,15 +34,15 @@ object WdlValues {
   // collection values
   sealed trait V_Collection extends V
   case class V_Pair(l: V, r: V) extends V_Collection
-  case class V_Array(value: Vector[V]) extends V_Collection
-  case class V_Map(value: SeqMap[V, V]) extends V_Collection
-  case class V_Struct(name: String, members: Map[String, V]) extends V_Collection
-  case class V_Object(members: Map[String, V]) extends V_Collection
+  case class V_Array(items: Vector[V]) extends V_Collection
+  case class V_Map(items: SeqMap[V, V]) extends V_Collection
+  case class V_Struct(name: String, fields: SeqMap[String, V]) extends V_Collection
+  case class V_Object(fields: SeqMap[String, V]) extends V_Collection
 
   // wrapper around another value to indicate it is associated with an
   // optional type
   case class V_Optional(value: V) extends V
 
   // results from calling a task or workflow
-  case class V_Call(name: String, members: Map[String, V]) extends V
+  case class V_Call(name: String, outputs: SeqMap[String, V]) extends V
 }